+++ src/methods/eventListener.js
@@ -1,0 +1,13 @@
+// var constant = require('./values/constants');
+// var pageElement = require('./values/elements');
+
+// module.exports = {
+//   showMoreReviews: function(pageNumber){
+//     if (isNextPageAvailable(pageNumber, constant.PAGE_SIZE)) {
+//       pageNumber++;
+//       renderReviews(pageNumber);
+//     } else {
+//       pageElement.moreReviewsBtn.classList.add('invisible');
+//     }
+//   }
+// }
