+++ src/values/constants.js
@@ -1,0 +1,4 @@
+module.exports = {
+	PAGE_SIZE: 3,
+	LOAD_TIMEOUT: 10000
+}
