+++ src/values/elements.js
@@ -5,5 +5,6 @@ module.exports = {
 	templateElement: document.querySelector('#review-template'),
 	reviewsNotFoundTemplate: document.querySelector('#reviews-not-found'),
 	reviewsNotFoundContainer: document.querySelector('.reviews-not-found'),
+	moreReviewsBtn: document.querySelector('.reviews-controls-more'),
 	moreReviewsBtn: document.querySelector('.reviews-controls-more')
 }
